==> script/01_explore_index_file.Rmd <==
---
title: "Exploring the Index File"
output: html_document
date: "2023-01-17"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, message = FALSE, warning = FALSE,
  cache = TRUE, cache.path = '.cache_knitr/explore_index_file/',
  fig.width = 7, fig.height = 4, dpi = 300,
  fig.path='figure/01_explore_index_file/'
)
```

In this document, we are utilising various metrics to understand the properties of the Lake Malawi cichlid pan-genome graph built using `minigraph` using default parameters. We aim to get an idea of the openness of the pangenome, how many sites of structural variation (SV) there are and how widespread/private alleles are.

## Setup
```{r, cache=FALSE}
library("readr")
library("ggplot2")
library("dplyr")
```

## Data import
```{r}
df_raw <- readr::read_delim(
  paste0(
    Sys.getenv("HOME"), 
    "/code/biospark/data/BIOGRID-ORCS-ALL-homo_sapiens-1.1.13.screens/BIOGRID-ORCS-SCREEN_INDEX-1.1.13.index.tab.txt"),
  col_types = paste0(
    "icccciific", # A-J
    "cccccccccc", # K-T
    "cccccccicc", # U-AD
    "cccfcccc"    # AE-AL
  ),
  na = c("", "NA", "-")
)
colnames(df_raw)[1] <- "SCREEN_ID"

# add broader category of cell types
metadata_celltype <- readr::read_csv(
  paste0(Sys.getenv("HOME"), "/code/biospark/data/metadata_celltype.csv"),
  col_types = "ccc", na = c("", "NA", "-") ) %>% 
  select(CELL_TYPE, CELL_TYPE_BROAD)

# add broader category for cancer cell lines
metadata_celltype_cancer <- readr::read_csv(
  paste0(Sys.getenv("HOME"), "/code/biospark/data/metadata_celltype_cancer.csv"),
  col_types = "ccc", na = c("", "NA", "-") ) %>% 
  select(CELL_LINE, CELL_TYPE_BROAD)


df <- bind_rows(
  filter(df_raw, CELL_TYPE != "Cancer Cell Line") %>% left_join(metadata_celltype, by='CELL_TYPE'),
  filter(df_raw, CELL_TYPE == "Cancer Cell Line") %>% left_join(distinct(metadata_celltype_cancer), by='CELL_LINE')
)

df
```

The following are useful columns to pay attention to:

More info: https://wiki.thebiogrid.org/doku.php/orcs:downloads:screen_index

##### Identifier
`2.SOURCE_ID`, `4.AUTHOR_NAME`, `5.SCREEN_NAME`

##### Quality
`6.SCORES_SIZE`, `7.FULL_SIZE`, `8.FULL_SIZE_AVAILABLE`, `9.NUMBER_OF_HITS`

```{r}
table(df$FULL_SIZE_AVAILABLE)
```

##### Computational analysis
`10.ANALYSIS`, `11.SIGNIFICANCE_INDICATOR`, `12.SIGNIFICANCE_CRITERIA`

```{r}
table(df$ANALYSIS)
```

##### Screen category
`13.THROUGHPUT`, `14.SCREEN_TYPE`, `15.SCREEN_FORMAT` (all 'Pool'), `16.EXPERIMENTAL_SETUP`

```{r}
table(df$EXPERIMENTAL_SETUP)
```


##### Library
`21.LIBRARY` (probably not useful), `22.LIBRARY_TYPE`, `23.LIBRARY_METHODOLOGY` (22 and 23 probably equivalent)

```{r}
table(df$LIBRARY_TYPE)
```

#### Cell
`25.CELL_LINE`, `26.CELL_TYPE`, `27.PHENOTYPE`

There are many cell lines, so I divided them into broader cateogories.
```{r}
table(df$CELL_TYPE_BROAD)
```

```{r}
table(df$PHENOTYPE)
```

There are `r length(table(df$CONDITION_NAME))` types of conditions.
```{r}
head(table(df$CONDITION_NAME), 12)
```

##### Other
`36.NOTES` (useful for humans to read)


## Exploration plots
```{r bar_librarymethod, fig.width=6, fig.height=3}
ggplot(df, aes(LIBRARY_METHODOLOGY) ) + 
  geom_histogram(stat='count') +
  ggtitle("Types of CRISPR screen") + 
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), vjust = -0.5, stat = "count") +
  scale_y_continuous(expand = expansion(mult = c(.05, .1))) +
  theme(aspect.ratio = 1)
```

Distribution of broad cell type categories.
```{r bar_celltypebroad, fig.width=7, fig.height=9}
ggplot(df, aes(CELL_TYPE_BROAD, fill = CELL_TYPE_BROAD) ) + 
  geom_histogram(stat='count') +
  ggtitle("Cell types in database") +
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), hjust = -0.2, stat = "count") +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(expand = expansion(mult = c(.05, .1))) +
  coord_flip() +
  theme(legend.position = 'none')
```

Distribution of cell types in the database. Note that "Cancer Cell Line" is excluded because they could come from multiple broader categories.
```{r bar_celltype, fig.width=10, fig.height=16}
cell_type_ordered <- df %>%
  filter(CELL_TYPE != "Cancer Cell Line") %>% 
  select(CELL_TYPE, CELL_TYPE_BROAD) %>% 
  distinct() %>% 
  arrange(CELL_TYPE_BROAD) %>% 
  .$CELL_TYPE

df %>%
  filter(CELL_TYPE != "Cancer Cell Line") %>% 
  ggplot(aes(factor(CELL_TYPE, levels = cell_type_ordered), fill = CELL_TYPE_BROAD) ) + 
  geom_histogram(stat='count') +
  ggtitle("Cell types in database") + 
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), hjust = -0.2, stat = "count") +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(expand = expansion(mult = c(.05, .1))) +
  coord_flip() +
  guides(fill = guide_legend(ncol = 1))
```

Phenotypes in database.
```{r bar_phenotype, fig.width=7, fig.height=7}
ggplot(df, aes(PHENOTYPE) ) + 
  geom_histogram(stat='count') +
  ggtitle("Phenotypes in database") + 
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), hjust = -0.2, stat = "count") +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(expand = expansion(mult = c(.05, .1))) +
  coord_flip()
```

This shows the distribution of cell types across the major phenotypes. In transparent are samples which have a `Timecourse` setup.
```{r facet_distribution, fig.width=12, fig.height=15}
df %>% 
  filter(PHENOTYPE %in% c("cell proliferation", "protein/peptide accumulation", "response to chemicals", "response to virus") ) %>% 
  ggplot(aes(CELL_TYPE_BROAD, fill = CELL_TYPE_BROAD) ) +
  geom_histogram(aes(alpha = EXPERIMENTAL_SETUP != "Timecourse"), color='black', stat='count')  +
  facet_wrap(.~PHENOTYPE) +
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), hjust = -0.2, stat = "count") +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(expand = expansion(mult = c(.05, .15))) +
  scale_alpha_manual(values=c(1, 0.35), breaks=c(T,F)) +
  ggtitle("Cell type distribution across main phenotypes, transparent: timecourse experiments") +
  coord_flip() +
  theme_classic() +
  theme(legend.position = 'none')
```

I also notice that most screens with cell proliferation as a phenotype are simply timecourses, with nothing in the experimental condition column. This suggests that they were simply observing cell growth.
```{r}
df %>% 
  filter(PHENOTYPE == "cell proliferation") %>% 
  .$EXPERIMENTAL_SETUP %>% table()

df %>% 
  filter(PHENOTYPE == "cell proliferation") %>% 
  .$CONDITION_NAME %>% is.na() %>% 
  table()
```

## More plots with just cell proliferation
We decide that we will focus only on samples which are annotated with a `cell proliferation` phenotype.

```{r}
df_filtered <- filter(df, PHENOTYPE == "cell proliferation")
```

Experiment types:
```{r subset-bar_librarymethod, fig.width=6, fig.height=3}
ggplot(df_filtered, aes(LIBRARY_METHODOLOGY) ) + 
  geom_histogram(stat='count') +
  ggtitle("Types of CRISPR screen") + 
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), vjust = -0.5, stat = "count") +
  scale_y_continuous(expand = expansion(mult = c(.05, .1))) +
  theme(aspect.ratio = 1)
```

Distribution of broad cell types in this subset.
```{r subset-bar_celltypebroad, fig.width=7, fig.height=9}
ggplot(df_filtered, aes(CELL_TYPE_BROAD, fill = CELL_TYPE_BROAD) ) + 
  geom_histogram(stat='count') +
  ggtitle("Cell types in database (cell proliferation)") +
  xlab(NULL) +  ylab("frequency") +
  geom_text(aes(label = ..count..), hjust = -0.2, stat = "count") +
  scale_x_discrete(limits = rev) +
  scale_y_continuous(expand = expansion(mult = c(.05, .1))) +
  coord_flip() +
  theme(legend.position = 'none')
```

Write filtered data frame focusing only on the "cell proliferation" screens.
```{r}
write_csv(df_filtered, paste0(Sys.getenv("HOME"), "/code/biospark/data/index_file_polished.csv") )
```
